Refactor favorites repository and use in viewmodel
