+++ app/src/main/java/com/ioannapergamali/mysmartroute/repository/FavoritesRepository.kt
@@ -2,7 +2,7 @@ package com.ioannapergamali.mysmartroute.repository
 
 
 import com.google.firebase.auth.FirebaseAuth
-
+import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.ioannapergamali.mysmartroute.data.local.PoIEntity
@@ -26,14 +26,37 @@ class FavoritesRepository {
         .collection("pois")
 
     /**
-     * Αποθηκεύει ή ενημερώνει ένα αγαπημένο POI με αναφορά στο έγγραφο του.
+     * Αποθηκεύει ή ενημερώνει ένα αγαπημένο POI.
      */
-    suspend fun saveFavorite(poi: PoIEntity) {
+    suspend fun saveFavorite(poiId: String) {
         val uid = userId()
         if (uid.isBlank()) return
-        val poiRef = firestore.collection("pois").document(poi.id)
-        userFavorites(uid).document(poi.id).set(mapOf("poiRef" to poiRef)).await()
+        val poiRef = firestore.collection("pois").document(poiId)
+        userFavorites(uid).document(poiId).set(mapOf("poiRef" to poiRef)).await()
+    }
 
+    /**
+     * Overload που δέχεται οντότητα POI.
+     */
+    suspend fun saveFavorite(poi: PoIEntity) = saveFavorite(poi.id)
+
+    /**
+     * Αφαιρεί ένα POI από τα αγαπημένα.
+     */
+    suspend fun removeFavorite(poiId: String) {
+        val uid = userId()
+        if (uid.isBlank()) return
+        userFavorites(uid).document(poiId).delete().await()
+    }
+
+    /**
+     * Επιστρέφει τις αναφορές όλων των αγαπημένων POIs.
+     */
+    suspend fun getFavoriteRefs(): List<DocumentReference> {
+        val uid = userId()
+        if (uid.isBlank()) return emptyList()
+        val snap = userFavorites(uid).get().await()
+        return snap.documents.mapNotNull { it.getDocumentReference("poiRef") }
     }
 }
